# pmaxis/services/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PMAxis API",
    "description": "Real-time prediction market data platform. Provides market data, prices, orderbooks, trades, signals, and analytics from Polymarket.\n\n**Rate limit:** 60 requests/minute per IP on all `/v1/*` routes.\n\n**Debug endpoints** require `X-Debug-Token` header (set `DEBUG_TOKEN` env var on the server).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://167.233.97.217:8088",
      "description": "Production VPS"
    },
    {
      "url": "http://localhost:8088",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/v1/markets": {
      "get": {
        "tags": ["Markets"],
        "summary": "List markets",
        "description": "Returns paginated markets ranked by recent activity. Supports filtering by status, category, tag, and series. Use `cursor` for pagination.",
        "operationId": "getMarkets",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "maximum": 200 }, "description": "Number of results to return" },
          { "name": "cursor", "in": "query", "schema": { "type": "string" }, "description": "Pagination cursor from previous response `next_cursor`" },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 }, "description": "Offset (use cursor instead for reliable pagination)" },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["ACTIVE", "CLOSED", "RESOLVED"] }, "description": "Filter by market status" },
          { "name": "category", "in": "query", "schema": { "type": "string", "example": "politics" }, "description": "Filter by category slug" },
          { "name": "tag", "in": "query", "schema": { "type": "string", "example": "trump" }, "description": "Filter by tag slug" },
          { "name": "series", "in": "query", "schema": { "type": "string", "example": "England" }, "description": "Filter by series name" },
          { "name": "exchange", "in": "query", "schema": { "type": "string", "default": "polymarket" }, "description": "Exchange source" },
          { "name": "source", "in": "query", "schema": { "type": "string", "enum": ["local", "gamma"] }, "description": "Use `gamma` to proxy directly to Polymarket Gamma API" }
        ],
        "responses": {
          "200": {
            "description": "Paginated market list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Market" } },
                    "next_cursor": { "type": "string" },
                    "has_more": { "type": "boolean" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/markets/search": {
      "get": {
        "tags": ["Markets"],
        "summary": "Search markets",
        "description": "Full-text search on market title and question. Case-insensitive substring match.",
        "operationId": "searchMarkets",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string", "example": "bitcoin" }, "description": "Search query" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["ACTIVE", "CLOSED", "RESOLVED"] } },
          { "name": "exchange", "in": "query", "schema": { "type": "string", "default": "polymarket" } }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Market" } }
              }
            }
          }
        }
      }
    },
    "/v1/markets/top": {
      "get": {
        "tags": ["Markets"],
        "summary": "Top markets by volume or trades",
        "description": "Returns the highest-activity markets over a time window, ranked by volume or trade count.",
        "operationId": "getTopMarkets",
        "parameters": [
          { "name": "by", "in": "query", "schema": { "type": "string", "enum": ["volume", "trades"], "default": "volume" }, "description": "Rank by volume (USDC) or trade count" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10, "maximum": 50 } },
          { "name": "period", "in": "query", "schema": { "type": "string", "enum": ["1h", "24h", "7d"], "default": "24h" }, "description": "Time window for aggregation" }
        ],
        "responses": {
          "200": {
            "description": "Top markets",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Market" } }
              }
            }
          }
        }
      }
    },
    "/v1/markets/{id}": {
      "get": {
        "tags": ["Markets"],
        "summary": "Get market detail",
        "description": "Full market detail including live price, orderbook state, and embedded 24h stats.",
        "operationId": "getMarket",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "example": "558935" }, "description": "Market ID" }
        ],
        "responses": {
          "200": { "description": "Market detail", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MarketDetail" } } } },
          "404": { "description": "Market not found" }
        }
      }
    },
    "/v1/markets/{id}/price": {
      "get": {
        "tags": ["Prices"],
        "summary": "Get live price for a market",
        "description": "Returns the latest mid-price, bid, ask, and spread from Redis (updated in real-time by the processor hub).",
        "operationId": "getPrice",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "example": "558935" } }
        ],
        "responses": {
          "200": {
            "description": "Live price",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Price" }
              }
            }
          },
          "404": { "description": "Price not found — market may not be actively traded" }
        }
      }
    },
    "/v1/prices": {
      "get": {
        "tags": ["Prices"],
        "summary": "Batch price lookup",
        "description": "Returns live prices for up to 100 market IDs in a single request. Returns `null` for markets with no live price.",
        "operationId": "getBatchPrices",
        "parameters": [
          { "name": "ids", "in": "query", "required": true, "schema": { "type": "string", "example": "558935,540817,540818" }, "description": "Comma-separated market IDs (max 100)" }
        ],
        "responses": {
          "200": {
            "description": "Map of market_id → price object (null if not found)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": { "$ref": "#/components/schemas/Price" }
                }
              }
            }
          }
        }
      }
    },
    "/v1/markets/{id}/orderbook": {
      "get": {
        "tags": ["Orderbook"],
        "summary": "Get live orderbook",
        "description": "Returns the current L2 orderbook state from Redis (updated in real-time).",
        "operationId": "getOrderbook",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "example": "558935" } }
        ],
        "responses": {
          "200": { "description": "Orderbook with bids and asks" },
          "404": { "description": "Orderbook not found" }
        }
      }
    },
    "/v1/markets/{id}/trades": {
      "get": {
        "tags": ["Trades"],
        "summary": "Get trade history for a market",
        "description": "Returns normalized trade events from ClickHouse. Use `source=data-api` to proxy live trades directly from Polymarket.",
        "operationId": "getMarketTrades",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "example": "558935" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100, "maximum": 5000 } },
          { "name": "from", "in": "query", "schema": { "type": "string", "example": "2026-06-01T00:00:00Z" }, "description": "Start time (RFC3339 or Unix ms)" },
          { "name": "to", "in": "query", "schema": { "type": "string" }, "description": "End time (RFC3339 or Unix ms)" },
          { "name": "source", "in": "query", "schema": { "type": "string", "enum": ["local", "data-api"] }, "description": "Use `data-api` to proxy directly from Polymarket" },
          { "name": "wallet", "in": "query", "schema": { "type": "string" }, "description": "Filter by wallet address (requires source=data-api)" }
        ],
        "responses": {
          "200": {
            "description": "Trade list",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Trade" } }
              }
            }
          }
        }
      }
    },
    "/v1/markets/{id}/price-history": {
      "get": {
        "tags": ["Price History"],
        "summary": "Price history for a market",
        "description": "Returns OHLCV data at the requested resolution from ClickHouse.",
        "operationId": "getMarketPriceHistory",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "example": "558935" } },
          { "name": "resolution", "in": "query", "schema": { "type": "string", "enum": ["1s", "1m", "5m", "1h"], "default": "1m" }, "description": "Candle resolution" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 500, "maximum": 20000 } },
          { "name": "from", "in": "query", "schema": { "type": "string" }, "description": "Start time (RFC3339 or Unix ms)" },
          { "name": "to", "in": "query", "schema": { "type": "string" }, "description": "End time (RFC3339 or Unix ms)" }
        ],
        "responses": {
          "200": { "description": "Price history array" }
        }
      }
    },
    "/v1/markets/{id}/candles": {
      "get": {
        "tags": ["Price History"],
        "summary": "OHLCV candles (TradingView UDF format)",
        "description": "Returns candles compatible with TradingView's UDF datafeed format. `from` and `to` are Unix seconds.",
        "operationId": "getCandles",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "example": "558935" } },
          { "name": "resolution", "in": "query", "schema": { "type": "string", "enum": ["1", "60"], "default": "1" }, "description": "1 = 1-minute bars, 60 = 1-hour bars" },
          { "name": "from", "in": "query", "schema": { "type": "integer", "example": 1704067200 }, "description": "Start time as Unix seconds" },
          { "name": "to", "in": "query", "schema": { "type": "integer", "example": 1704153600 }, "description": "End time as Unix seconds" }
        ],
        "responses": {
          "200": { "description": "TradingView UDF candle response" }
        }
      }
    },
    "/v1/markets/{id}/summary": {
      "get": {
        "tags": ["Markets"],
        "summary": "Market summary",
        "description": "Unified view of a market including live price, latest signals, and 24h stats in one response.",
        "operationId": "getMarketSummary",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "example": "558935" } }
        ],
        "responses": {
          "200": { "description": "Market summary" }
        }
      }
    },
    "/v1/markets/{id}/stats": {
      "get": {
        "tags": ["Markets"],
        "summary": "Market 24h stats",
        "description": "Returns 24h OHLCV stats and price change % for a market.",
        "operationId": "getMarketStats",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "example": "558935" } }
        ],
        "responses": {
          "200": {
            "description": "24h stats",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MarketStats" }
              }
            }
          }
        }
      }
    },
    "/v1/markets/{id}/related": {
      "get": {
        "tags": ["Markets"],
        "summary": "Related markets (same event group)",
        "description": "Returns sibling markets that share the same `event_id`. Useful for multi-outcome events (e.g. all team outcomes in a tournament).",
        "operationId": "getRelatedMarkets",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "example": "558935" } }
        ],
        "responses": {
          "200": { "description": "Related markets" }
        }
      }
    },
    "/v1/markets/{id}/positions": {
      "get": {
        "tags": ["Positions"],
        "summary": "Positions for a market",
        "description": "Returns all open positions for a specific market. Proxied from Polymarket Data API.",
        "operationId": "getMarketPositions",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "example": "558935" } },
          { "name": "wallet", "in": "query", "schema": { "type": "string" }, "description": "Filter by wallet address" }
        ],
        "responses": {
          "200": { "description": "Positions" }
        }
      }
    },
    "/v1/markets/{id}/signals": {
      "get": {
        "tags": ["Signals"],
        "summary": "Trading signals for a market",
        "description": "Returns the latest trading signals generated by the processor hub for a specific market.",
        "operationId": "getSignalsByMarket",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "example": "558935" } },
          { "name": "type", "in": "query", "schema": { "type": "string" }, "description": "Filter by signal type (e.g. momentum, volume_spike)" }
        ],
        "responses": {
          "200": { "description": "Signals array" }
        }
      }
    },
    "/v1/events": {
      "get": {
        "tags": ["Taxonomy"],
        "summary": "List events",
        "description": "Proxied directly from Polymarket Gamma API `/events`. Supports all Gamma API query params.",
        "operationId": "getEvents",
        "parameters": [
          { "name": "exchange", "in": "query", "schema": { "type": "string", "default": "polymarket" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "active", "in": "query", "schema": { "type": "boolean" } },
          { "name": "closed", "in": "query", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": { "description": "Events from Polymarket Gamma API" }
        }
      }
    },
...
    "/v1/positions": {
      "get": {
        "tags": ["Positions"],
        "summary": "Open positions for a wallet",
        "description": "Returns all open (unsettled) positions for a Polymarket wallet address. Pass the full 0x address in the `wallet` param. Proxied from Polymarket Data API.",
        "operationId": "getPositions",
        "parameters": [
          { "name": "wallet", "in": "query", "required": true, "schema": { "type": "string", "example": "0xAdA100Db00Ca00073811820692005400218FcE1f" }, "description": "Full Ethereum wallet address (required)" }
        ],
        "responses": {
          "200": { "description": "Open positions for the wallet" },
          "400": { "description": "wallet is required" }
        }
      }
    },
    "/v1/positions/closed": {
      "get": {
        "tags": ["Positions"],
        "summary": "Closed positions for a wallet",
        "description": "Returns all closed/settled positions for a Polymarket wallet address. Pass the full 0x address in the `wallet` param. Proxied from Polymarket Data API.",
        "operationId": "getClosedPositions",
        "parameters": [
          { "name": "wallet", "in": "query", "required": true, "schema": { "type": "string", "example": "0xAdA100Db00Ca00073811820692005400218FcE1f" }, "description": "Full Ethereum wallet address (required)" }
        ],
        "responses": {
          "200": { "description": "Closed positions for the wallet" },
          "400": { "description": "wallet is required" }
        }
      }
    },
    "/v1/wallets/watch": {
      "post": {
        "tags": ["Wallet Activity"],
        "summary": "Watch a wallet",
        "description": "Adds a wallet address to the watch list. Once watched, all on-chain ERC-1155 settlement trades involving this address will be captured in real-time and stored in `wallet_activity`. Also triggers live WebSocket alerts via `subscribe_wallet`.",
        "operationId": "watchWallet",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["address"],
                "properties": {
                  "address": { "type": "string", "example": "0xE111180000d2663C0091e4f400237545B87B996B", "description": "Ethereum wallet address to watch" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Wallet added to watch list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "address": { "type": "string" },
                    "watching": { "type": "boolean", "example": true }
                  }
                }
              }
            }
          },
          "400": { "description": "address is required" }
        }
      }
    },
...
    "/v1/wallets/{address}/watch": {
      "delete": {
        "tags": ["Wallet Activity"],
        "summary": "Unwatch a wallet",
        "description": "Removes a wallet address from the watch list. New trades will no longer be captured, but existing `wallet_activity` history is retained until the 90-day TTL expires.",
        "operationId": "unwatchWallet",
        "parameters": [
          { "name": "address", "in": "path", "required": true, "schema": { "type": "string", "example": "0xE111180000d2663C0091e4f400237545B87B996B" }, "description": "Wallet address to unwatch" }
        ],
        "responses": {
          "200": {
            "description": "Wallet removed from watch list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "address": { "type": "string" },
                    "watching": { "type": "boolean", "example": false }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/wallets/{address}/activity": {
      "get": {
        "tags": ["Wallet Activity"],
        "summary": "Wallet activity history (watched only)",
        "description": "Returns stored on-chain trade activity for a **watched** wallet from ClickHouse. Data is only captured from the moment the wallet was added to the watch list. For any wallet without pre-watching, use `/onchain` instead.",
        "operationId": "getWalletActivity",
        "parameters": [
          { "name": "address", "in": "path", "required": true, "schema": { "type": "string", "example": "0xE111180000d2663C0091e4f400237545B87B996B" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "maximum": 200 } }
        ],
        "responses": {
          "200": {
            "description": "Wallet activity records",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WalletActivityResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/wallets/{address}/onchain": {
      "get": {
        "tags": ["Wallet Activity"],
        "summary": "On-chain trades for any wallet",
        "description": "Returns raw on-chain ERC-1155 settlement trades for **any** wallet address (maker OR taker), directly from the `onchain_trades` table. No pre-watching required — works retroactively for trades within the 14-day retention window.",
        "operationId": "getWalletOnchain",
        "parameters": [
          { "name": "address", "in": "path", "required": true, "schema": { "type": "string", "example": "0xE111180000d2663C0091e4f400237545B87B996B" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "maximum": 200 } }
        ],
        "responses": {
          "200": {
            "description": "On-chain trade records",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WalletOnchainResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/wallets/{address}/summary": {
      "get": {
        "tags": ["Wallet Activity"],
        "summary": "Wallet trading summary",
        "description": "Returns aggregated trading stats for a watched wallet: total trades, total volume, buy/sell breakdown, and first/last seen timestamps. Queries the `wallet_activity` table.",
        "operationId": "getWalletSummary",
        "parameters": [
          { "name": "address", "in": "path", "required": true, "schema": { "type": "string", "example": "0xE111180000d2663C0091e4f400237545B87B996B" } }
        ],
        "responses": {
          "200": {
            "description": "Wallet summary stats",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WalletSummary" }
              }
            }
          }
        }
      }
    },
    "/debug/markets": {
      "get": {
        "tags": ["Debug"],
        "summary": "All markets in Redis",
        "description": "Returns raw Redis state for all tracked markets. Requires `X-Debug-Token` header.",
        "operationId": "getDebugMarkets",
        "security": [{ "DebugToken": [] }],
        "responses": {
          "200": { "description": "Markets from Redis" },
          "403": { "description": "debug endpoints disabled" }
        }
      }
    },
    "/debug/markets/{id}": {
      "get": {
        "tags": ["Debug"],
        "summary": "Single market Redis state",
        "operationId": "getDebugMarket",
        "security": [{ "DebugToken": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Market Redis state" },
          "403": { "description": "debug endpoints disabled" }
        }
      }
    },
    "/debug/trades": {
      "get": {
        "tags": ["Debug"],
        "summary": "Recent trades from Redis",
        "operationId": "getDebugTrades",
        "security": [{ "DebugToken": [] }],
        "responses": {
          "200": { "description": "Recent trades" },
          "403": { "description": "debug endpoints disabled" }
        }
      }
    },
    "/debug/orderbook/{id}": {
      "get": {
        "tags": ["Debug"],
        "summary": "Raw orderbook from Redis",
        "operationId": "getDebugOrderbook",
        "security": [{ "DebugToken": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Raw orderbook" },
          "403": { "description": "debug endpoints disabled" }
        }
      }
    },
    "/debug/signals": {
      "get": {
        "tags": ["Debug"],
        "summary": "All signals from Redis",
        "operationId": "getDebugSignals",
        "security": [{ "DebugToken": [] }],
        "responses": {
          "200": { "description": "All signals" },
          "403": { "description": "debug endpoints disabled" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "DebugToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Debug-Token",
        "description": "Set the `DEBUG_TOKEN` environment variable on the server, then pass that value here to access debug endpoints."
      }
    },
    "schemas": {
...
      "MarketDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/Market" },
          {
            "type": "object",
            "properties": {
              "condition_id": { "type": "string" },
              "stats_24h": { "$ref": "#/components/schemas/MarketStats" }
            }
          }
        ]
      },
      "Price": {
        "type": "object",
        "properties": {
          "market_id": { "type": "string" },
          "price": { "type": "number", "format": "float" },
          "best_bid": { "type": "number", "format": "float" },
          "best_ask": { "type": "number", "format": "float" },
          "mid_price": { "type": "number", "format": "float" },
          "spread": { "type": "number", "format": "float" },
          "volume_24h": { "type": "number", "format": "float" },
          "timestamp": { "type": "integer", "description": "Unix milliseconds" }
        }
      },
...
    }
  }
}